backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.34.3
pydantic-settings==2.10.1
python-dotenv==1.1.1
httpx==0.28.1
numpy==1.26.4
pypdf==5.4.0
python-docx==1.1.2
faiss-cpu==1.8.0.post1
pyinstaller==6.15.0
